Generate ts types for cql
