+++ .gitignore
@@ -4,3 +4,4 @@ dev/pki/*
 !dev/pki/pki
 Cargo.lock
 .vscode/*
+pkg/
