+++ libs/focus_api/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "focus_api"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+validated = "0.4.0"
+nonempty-collections = "0.1.4"
+tsify = "0.4.5"
+wasm-bindgen = "0.2.89"
+serde = { version = "1.0.152", features = ["serde_derive"] }
+chrono = "0.4.31"
+
+[dev-dependencies]
+pretty_assertions = "1.4.0"
+tokio-test = "0.4.2"
